pyproject.toml
[tool.poetry]
name = "biodegradablehumiliatingrelationaldatabase"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
PyNaCl = "^1.4.0"
youtube_dl = "^2021.6.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
